refactor: Refactor package structure
