+++ src/main/java/bowling/domain/Player.java
@@ -1,5 +1,7 @@
 package bowling.domain;
 
+import bowling.domain.frame.Frame;
+
 import java.util.List;
 
 public class Player {
