+++ src/test/java/bowling/domain/pitch/NormalTest.java
@@ -1,5 +1,6 @@
-package bowling.domain;
+package bowling.domain.pitch;
 
+import bowling.domain.KnockedPins;
 import org.junit.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
